Move integration test package
