+++ test/integration/selfhostedshootexposure/selfhostedshootexposure_test.go
@@ -1,4 +1,4 @@
-package stackit
+package selfhostedshootexposure
 
 import (
 	"bytes"
